alloy_provider::bindings

Module IMulticall3

Multicall3 bindings.

interface IMulticall3 {
    struct Call { address target; bytes callData; }
    struct Call3 { address target; bool allowFailure; bytes callData; }
    struct Call3Value { address target; bool allowFailure; uint256 value; bytes callData; }
    struct Result { bool success; bytes returnData; }
    function aggregate(Call[] calldata calls) external payable returns (uint256 blockNumber, bytes[] memory returnData);
    function aggregate3(Call3[] calldata calls) external payable returns (Result[] memory returnData);
    function aggregate3Value(Call3Value[] calldata calls) external payable returns (Result[] memory returnData);
    function blockAndAggregate(Call[] calldata calls) external payable returns (uint256 blockNumber, bytes32 blockHash, Result[] memory returnData);
    function getBasefee() external view returns (uint256 basefee);
    function getBlockHash(uint256 blockNumber) external view returns (bytes32 blockHash);
    function getBlockNumber() external view returns (uint256 blockNumber);
    function getChainId() external view returns (uint256 chainid);
    function getCurrentBlockCoinbase() external view returns (address coinbase);
    function getCurrentBlockDifficulty() external view returns (uint256 difficulty);
    function getCurrentBlockGasLimit() external view returns (uint256 gaslimit);
    function getCurrentBlockTimestamp() external view returns (uint256 timestamp);
    function getEthBalance(address addr) external view returns (uint256 balance);
    function getLastBlockHash() external view returns (bytes32 blockHash);
    function tryAggregate(bool requireSuccess, Call[] calldata calls) external payable returns (Result[] memory returnData);
    function tryBlockAndAggregate(bool requireSuccess, Call[] calldata calls) external payable returns (uint256 blockNumber, bytes32 blockHash, Result[] memory returnData);
}

Structs

Call

Call3

Call3Value

Result

aggregate3Call

Function with signature aggregate3((address,bool,bytes)[]) and selector 0x82ad56cb.

aggregate3Return

Container type for the return parameters of the aggregate3((address,bool,bytes)[]) function.

aggregate3ValueCall

Function with signature aggregate3Value((address,bool,uint256,bytes)[]) and selector 0x174dea71.

aggregate3ValueReturn

Container type for the return parameters of the aggregate3Value((address,bool,uint256,bytes)[]) function.

aggregateCall

Function with signature aggregate((address,bytes)[]) and selector 0x252dba42.

aggregateReturn

Container type for the return parameters of the aggregate((address,bytes)[]) function.

blockAndAggregateCall

Function with signature blockAndAggregate((address,bytes)[]) and selector 0xc3077fa9.

blockAndAggregateReturn

Container type for the return parameters of the blockAndAggregate((address,bytes)[]) function.

getBasefeeCall

Function with signature getBasefee() and selector 0x3e64a696.

getBasefeeReturn

Container type for the return parameters of the getBasefee() function.

getBlockHashCall

Function with signature getBlockHash(uint256) and selector 0xee82ac5e.

getBlockHashReturn

Container type for the return parameters of the getBlockHash(uint256) function.

getBlockNumberCall

Function with signature getBlockNumber() and selector 0x42cbb15c.

getBlockNumberReturn

Container type for the return parameters of the getBlockNumber() function.

getChainIdCall

Function with signature getChainId() and selector 0x3408e470.

getChainIdReturn

Container type for the return parameters of the getChainId() function.

getCurrentBlockCoinbaseCall

Function with signature getCurrentBlockCoinbase() and selector 0xa8b0574e.

getCurrentBlockCoinbaseReturn

Container type for the return parameters of the getCurrentBlockCoinbase() function.

getCurrentBlockDifficultyCall

Function with signature getCurrentBlockDifficulty() and selector 0x72425d9d.

getCurrentBlockDifficultyReturn

Container type for the return parameters of the getCurrentBlockDifficulty() function.

getCurrentBlockGasLimitCall

Function with signature getCurrentBlockGasLimit() and selector 0x86d516e8.

getCurrentBlockGasLimitReturn

Container type for the return parameters of the getCurrentBlockGasLimit() function.

getCurrentBlockTimestampCall

Function with signature getCurrentBlockTimestamp() and selector 0x0f28c97d.

getCurrentBlockTimestampReturn

Container type for the return parameters of the getCurrentBlockTimestamp() function.

getEthBalanceCall

Function with signature getEthBalance(address) and selector 0x4d2301cc.

getEthBalanceReturn

Container type for the return parameters of the getEthBalance(address) function.

getLastBlockHashCall

Function with signature getLastBlockHash() and selector 0x27e86d6e.

getLastBlockHashReturn

Container type for the return parameters of the getLastBlockHash() function.

tryAggregateCall

Function with signature tryAggregate(bool,(address,bytes)[]) and selector 0xbce38bd7.

tryAggregateReturn

Container type for the return parameters of the tryAggregate(bool,(address,bytes)[]) function.

tryBlockAndAggregateCall

Function with signature tryBlockAndAggregate(bool,(address,bytes)[]) and selector 0x399542e9.

tryBlockAndAggregateReturn

Container type for the return parameters of the tryBlockAndAggregate(bool,(address,bytes)[]) function.

Enums

IMulticall3Calls

Container for all the IMulticall3 function calls.

Statics

BYTECODE

The creation / init bytecode of the contract.