Trait alloy_provider::ext::TraceApi

pub trait TraceApi<N, T>: Send + Sync
where
    N: Network,
    T: Transport + Clone,
{
    // Required methods
    fn trace_call<'a, 'b>(
        &self,
        request: &'a N::TransactionRequest,
        trace_type: &'b [TraceType],
    ) -> RpcWithBlock<T, (&'a N::TransactionRequest, &'b [TraceType]), TraceResults>;
    fn trace_call_many<'a>(
        &self,
        request: TraceCallList<'a, N>,
    ) -> RpcWithBlock<T, TraceCallList<'a, N>, TraceResults>;
    fn trace_transaction<'life0, 'async_trait>(
        &'life0 self,
        hash: TxHash,
    ) -> Pin<Box<dyn Future<Output = TransportResult<Vec<LocalizedTransactionTrace>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn trace_block<'life0, 'async_trait>(
        &'life0 self,
        block: BlockNumberOrTag,
    ) -> Pin<Box<dyn Future<Output = TransportResult<Vec<LocalizedTransactionTrace>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn trace_replay_transaction<'life0, 'life1, 'async_trait>(
        &'life0 self,
        hash: TxHash,
        trace_type: &'life1 [TraceType],
    ) -> Pin<Box<dyn Future<Output = TransportResult<TraceResults>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
}

Available on crate feature trace-api only.

Trace namespace rpc interface that gives access to several non-standard RPC methods.

Required Methods

fn trace_call<'a, 'b>( &self, request: &'a N::TransactionRequest, trace_type: &'b [TraceType], ) -> RpcWithBlock<T, (&'a N::TransactionRequest, &'b [TraceType]), TraceResults>

Executes the given transaction and returns a number of possible traces.

Note

Not all nodes support this call.

fn trace_call_many<'a>( &self, request: TraceCallList<'a, N>, ) -> RpcWithBlock<T, TraceCallList<'a, N>, TraceResults>

Traces multiple transactions on top of the same block, i.e. transaction n will be executed on top of the given block with all n - 1 transaction applied first.

Allows tracing dependent transactions.

Note

Not all nodes support this call.

fn trace_transaction<'life0, 'async_trait>( &'life0 self, hash: TxHash, ) -> Pin<Box<dyn Future<Output = TransportResult<Vec<LocalizedTransactionTrace>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Parity trace transaction.

fn trace_block<'life0, 'async_trait>( &'life0 self, block: BlockNumberOrTag, ) -> Pin<Box<dyn Future<Output = TransportResult<Vec<LocalizedTransactionTrace>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Trace all transactions in the given block.

Note

Not all nodes support this call.

fn trace_replay_transaction<'life0, 'life1, 'async_trait>( &'life0 self, hash: TxHash, trace_type: &'life1 [TraceType], ) -> Pin<Box<dyn Future<Output = TransportResult<TraceResults>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Replays a transaction.