alloy_provider::fillers
Struct FillProvider

pub struct FillProvider<F, P, T, N>
where
    F: TxFiller<N>,
    P: Provider<T, N>,
    T: Transport + Clone,
    N: Network,
{ /* private fields */ }

A Provider that applies one or more TxFillers.

Fills arbitrary properties in a transaction request by composing multiple fill layers. This struct should always be the outermost layer in a provider stack, and this is enforced when using ProviderBuilder::filler to construct this layer.

Users should NOT use this struct directly. Instead, use ProviderBuilder::filler to construct and apply it to a stack.

Implementations

impl<F, P, T, N> FillProvider<F, P, T, N>
where F: TxFiller<N>, P: Provider<T, N>, T: Transport + Clone, N: Network,

pub fn new(inner: P, filler: F) -> Self

Creates a new FillProvider with the given filler and inner provider.

pub fn join_with<Other: TxFiller<N>>( self, other: Other, ) -> FillProvider<JoinFill<F, Other>, P, T, N>

Joins a filler to this provider

pub async fn fill( &self, tx: N::TransactionRequest, ) -> TransportResult<SendableTx<N>>

Fills the transaction request, using the configured fillers