Module alloy_provider::heart

Block heartbeat and pending transaction watcher.

Structs

Heartbeat 🔒

A heartbeat task that receives blocks and watches for transactions.

HeartbeatHandle 🔒

A handle to the heartbeat task.

PendingTransaction

Represents a transaction that is yet to be confirmed a specified number of times.

PendingTransactionBuilder

A builder for configuring a pending transaction watcher.

PendingTransactionConfig

Configuration for watching a pending transaction.

TxWatcher 🔒