Module alloy_provider::layers

Useful layer implementations for the provider. Currently this module contains the AnvilLayer, AnvilProvider and ChainLayer types.

Modules

anvil 🔒 anvil-node

chain 🔒

Structs

AnvilLayeranvil-node

A layer that wraps an Anvil config. The config will be used to spawn an AnvilInstance when the layer is applied, or when the user requests any information about the anvil node (e.g. via the AnvilLayer::ws_endpoint_url method ).

AnvilProvideranvil-node

A provider that wraps an AnvilInstance, preventing the instance from being dropped while the provider is in use.

ChainLayer

A layer that wraps a [NamedChain]. The layer will be used to set the client’s poll interval based on the average block time for this chain.