Module alloy_provider::layers::anvil

Available on crate feature anvil-node only.

Structs

AnvilLayer

A layer that wraps an Anvil config. The config will be used to spawn an AnvilInstance when the layer is applied, or when the user requests any information about the anvil node (e.g. via the AnvilLayer::ws_endpoint_url method ).

AnvilProvider

A provider that wraps an AnvilInstance, preventing the instance from being dropped while the provider is in use.