Struct alloy_provider::layers::CacheProvider

pub struct CacheProvider<P, T, N> { /* private fields */ }

The CacheProvider holds the underlying in-memory LRU cache and overrides methods from the Provider trait. It attempts to fetch from the cache and fallbacks to the RPC in case of a cache miss.

Most importantly, the CacheProvider adds save_cache and load_cache methods to the provider interface, allowing users to save the cache to disk and load it from there on demand.

Implementations

impl<P, T, N> CacheProvider<P, T, N>
where P: Provider<T, N>, T: Transport + Clone, N: Network,

pub const fn new(inner: P, cache: SharedCache) -> Self

Instantiate a new cache provider.