Module alloy_provider::provider::root

Structs

RootProvider

The root provider manages the RPC client and the heartbeat. It is at the base of every provider stack.

RootProviderInner 🔒

The root provider manages the RPC client and the heartbeat. It is at the base of every provider stack.

Functions

builder

Helper function to directly access ProviderBuilder with minimal generics.