Struct alloy_provider::ProviderBuilder

pub struct ProviderBuilder<L, N = Ethereum> {
    layer: L,
    network: PhantomData<N>,
}

A builder for constructing a Provider from various layers.

This type is similar to tower::ServiceBuilder, with extra complication around maintaining the network and transport types.

Fields
layer: Lnetwork: PhantomData<N>
Implementations

impl<N> ProviderBuilder<Identity, N>

pub const fn new() -> Self

Create a new ProviderBuilder.

impl<L, N> ProviderBuilder<L, N>

pub fn layer<Inner>(self, layer: Inner) -> ProviderBuilder<Stack<Inner, L>, N>

Add a layer to the stack being built. This is similar to tower::ServiceBuilder::layer.

Note:

Layers are added in outer-to-inner order, as in tower::ServiceBuilder. The first layer added will be the first to see the request.

pub fn signer<S>( self, signer: S ) -> ProviderBuilder<Stack<SignerLayer<S>, L>, N>

Add a signer layer to the stack being built.

See SignerLayer.

pub fn with_gas_estimation( self ) -> ProviderBuilder<Stack<GasEstimatorLayer, L>, N>

Add gas estimation to the stack being built.

See GasEstimatorLayer

pub fn with_nonce_management( self ) -> ProviderBuilder<Stack<NonceManagerLayer, L>, N>

Add nonce management to the stack being built.

See NonceManagerLayer

pub fn with_recommended_layers( self ) -> ProviderBuilder<Stack<NonceManagerLayer, Stack<GasEstimatorLayer, L>>, N>

Add preconfigured set of layers handling gas estimation and nonce management

pub fn network<Net: Network>(self) -> ProviderBuilder<L, Net>

Change the network.

By default, the network is Ethereum. This method must be called to configure a different network.

builder.network::<Arbitrum>()

pub fn provider<P, T>(self, provider: P) -> L::Provider
where L: ProviderLayer<P, N, T>, P: Provider<N, T>, T: Transport + Clone, N: Network,

Finish the layer stack by providing a root Provider, outputting the final Provider type with all stack components.

pub fn on_client<T>(self, client: RpcClient<T>) -> L::Provider
where L: ProviderLayer<RootProvider<N, T>, N, T>, T: Transport + Clone, N: Network,

Finish the layer stack by providing a root RpcClient, outputting the final Provider type with all stack components.

This is a convenience function for ProviderBuilder::provider<RpcClient>.

pub async fn on_builtin(self, s: &str) -> Result<L::Provider, TransportError>
where L: ProviderLayer<RootProvider<N, BoxTransport>, N, BoxTransport>, N: Network,

Finish the layer stack by providing a connection string for a built-in transport type, outputting the final Provider type with all stack components.

This is a convenience function for

pub async fn on_ws( self, connect: WsConnect ) -> Result<L::Provider, TransportError>
where L: ProviderLayer<RootProvider<N, PubSubFrontend>, N, PubSubFrontend>, N: Network,

Available on crate feature ws only.

Build this provider with a websocket connection.

pub async fn on_ipc<T>( self, connect: IpcConnect<T> ) -> Result<L::Provider, TransportError>
where IpcConnect<T>: PubSubConnect, L: ProviderLayer<RootProvider<N, PubSubFrontend>, N, PubSubFrontend>, N: Network,

Available on crate feature ipc only.

Build this provider with an IPC connection.