alloy_provider

Trait Caller

pub trait Caller<N, Resp>: Send + Sync
where
    N: Network,
    Resp: RpcRecv,
{
    // Required methods
    fn call(
        &self,
        params: EthCallParams<N>,
    ) -> TransportResult<ProviderCall<EthCallParams<N>, Resp>>;
    fn estimate_gas(
        &self,
        params: EthCallParams<N>,
    ) -> TransportResult<ProviderCall<EthCallParams<N>, Resp>>;
    fn call_many(
        &self,
        params: EthCallManyParams<'_>,
    ) -> TransportResult<ProviderCall<EthCallManyParams<'static>, Resp>>;
}

Trait that helpes convert EthCall into a ProviderCall.

Required Methods

fn call( &self, params: EthCallParams<N>, ) -> TransportResult<ProviderCall<EthCallParams<N>, Resp>>

Method that needs to be implemented to convert to a ProviderCall.

This method sends the request to relevant data source and returns a ProviderCall.

fn estimate_gas( &self, params: EthCallParams<N>, ) -> TransportResult<ProviderCall<EthCallParams<N>, Resp>>

Method that needs to be implemented for estimating gas using “eth_estimateGas” for the transaction.

fn call_many( &self, params: EthCallManyParams<'_>, ) -> TransportResult<ProviderCall<EthCallManyParams<'static>, Resp>>

Method that needs to be implemented for "eth_callMany" RPC requests.