Crate alloy_provider

alloy-provider

Interface with an Ethereum blockchain.

This crate contains the Provider trait, which exposes Ethereum JSON-RPC methods. Providers in alloy are similar to ethers.js providers. They manage an RpcClient and allow other parts of the program to easily make RPC calls.

Unlike an ethers.js Provider, an alloy Provider is network-aware. It is parameterized with a Network from [alloy-networks]. This allows the Provider to expose a consistent interface to the rest of the program, while adjusting request and response types to match the underlying blockchain.

Providers can be composed via stacking. For example, a Provider that tracks the nonce for a given address can be stacked onto a Provider that signs transactions to create a Provider that can send signed transactions with correct nonces.

The ProviderBuilder struct can quickly create a stacked provider, similar to tower::ServiceBuilder.

Feature flags

pubsub - Enable support for subscription methods.

ws - Enable WebSocket support. Implictly enables pubsub.

ipc - Enable IPC support. Implictly enables pubsub.

Usage

TODO :)

Re-exports

pub use provider::SendableTx;

pub use alloy_network as network;

pub use alloy_network::Network;

Modules

builder 🔒

chain 🔒

ext

Extended APIs for the provider module.

fillers

Transaction Fillers

heart 🔒

Block heartbeat and pending transaction watcher.

layers

Useful layer implementations for the provider. Currently this module contains the AnvilLayer and AnvilProvider types, when the anvil feature is enabled.

provider 🔒

utils

Provider-related utilities.

Structs

Identity

An identity layer that does nothing.

PendingTransaction

Represents a transaction that is yet to be confirmed a specified number of times.

PendingTransactionBuilder

A builder for configuring a pending transaction watcher.

PendingTransactionConfig

Configuration for watching a pending transaction.

ProviderBuilder

A builder for constructing a Provider from various layers.

RootProvider

The root provider manages the RPC client and the heartbeat. It is at the base of every provider stack.

Stack

A stack of two providers.

Traits

Provider

Provider is parameterized with a network and a transport. The default transport is type-erased, but you can do Provider<Http, N>.

ProviderLayer

A layering abstraction in the vein of tower::Layer

WalletProvider

Trait for Providers, Fill stacks, etc, which contain [NetworkSigner].

Type Aliases

FilterPollerBuilder

A task that polls the provider with eth_getFilterChanges, returning a list of R.

HyperProviderhyper

Type alias for a RootProvider using the Http transport and a hyper client.

ReqwestProviderreqwest

Type alias for a RootProvider using the Http transport and a reqwest client.