Module alloy_provider::layers

Provider layers.

Layers decorate a Provider, transforming various inputs and outputs of the root provider, depending on the layers used.

Modules

gas 🔒

nonce 🔒

signer 🔒

Structs

GasEstimatorLayer

A layer that populates gas related fields in transaction requests if unset.

GasEstimatorProvider

A provider that estimates gas for transactions.

ManagedNonceProvider

A provider that manages account nonces.

NonceManagerLayer

A layer that fills nonces on transactions.

SignerLayer

A layer that signs transactions locally.

SignerProvider

A locally-signing provider.